requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
crayons==0.1.2
Django==2.0.5
django-appconf==1.0.2
django-bootstrap-form==3.4
django-filter==2.0.0
django-ipware==1.1.6
django-user-accounts==2.0.3
djangorestframework==3.9.0
idna==2.8
jsonfield==2.0.2
pinax-cli==1.1.4
pinax-eventlog==2.0.3
pinax-stripe==4.0.0
pinax-templates==2.0.1
pinax-webanalytics==4.0.2
pytz==2018.9
requests==2.21.0
six==1.12.0
stripe==2.17.0
urllib3==1.24.1
